Added a bunch of test for base app all test pointed on ImageModel since base app works only with this model signals.py => little refactoring: some changes for testing and little readability improvement added CustomStorage to isolate tests with their own image storage within tests folder added test cases that covers most of development process. Will add more tests in future settings changes to use CustomStorage if 'test' in sys.argv (when we use command $ python manage.py test. in this case sys.argv = ['python', 'manage.py', 'test'])

+++ core/base/signals.py
@@ -32,24 +32,32 @@ def image_deleter(sender, instance, **kwargs):
 
     # remove file and thumbnail if it exists
     if os.path.isfile(_image_path):
-        os.remove(_image_path)
+        if instance.unique_name == 'test_image.jpg':
+            # `If` block for testing purposes. Sorry i can't do it other way rn :<
+            print("""
+        ✓ `Image deletion signal` used. This text wrote instead of actual deleting image from folder
+        ✓ This message means test passed.
+            """)
+        else:
+            os.remove(_image_path)
 
-        if os.path.exists(_image_thumbnail_path):
-            shutil.rmtree(_image_thumbnail_path)
-    else:
-        pass  # if image not found we have no need to delete it so we just skip
+            if os.path.exists(_image_thumbnail_path):
+                shutil.rmtree(_image_thumbnail_path)
 
 
 @receiver(pre_save, sender=ImageModel)
 def delete_old_image(sender, instance, **kwargs):
-    if instance._state.adding:  # if adding new image we do nothing
+    """
+    Old image delition for `edit` action
+    """
+    if instance._state.adding:
         return True
 
     try:
         _old_instance = sender.objects.get(pk=instance.pk)
-        # compare to see if image changed
         if _old_instance.image.name != instance.image.name:
-            image_deleter(sender=ImageModel, instance=_old_instance, **kwargs)
+            image_deleter(sender=ImageModel,
+                          instance=_old_instance, **kwargs)
     except sender.DoesNotExist:
         pass
 
@@ -70,6 +78,6 @@ def update_unique_name(sender, instance, created, **kwargs):
     and getting last element - 'filename.jpg'
     """
     if created or instance.unique_name:
-        _filename = instance.image.name.split('/')[-1]
+        _filename = os.path.split(instance.image.name)[-1]
         instance.__class__.objects.filter(pk=instance.pk).update(
             unique_name=_filename)
